move specs to test dir
