+++ test/http/views/card.hbs
@@ -1,0 +1,66 @@
+<div id="card-image">
+    <img src="{{this.image}}" alt="{{this.name}} - ({{this.set.code}})"/>
+</div>
+
+<div id="card-identity">
+    <div id="card-name">{{this.name}}</div>
+    <div id="card-set">
+        <i class="ss ss-{{this.keyruneCode}} ss-2x"></i>
+        <span id="set-name">{{this.set.name}}</span>
+        <span id="set-code"> ({{this.set.code}})</span>
+    </div>
+</div>
+
+<div id="play-info">
+    <h4>
+        <span>{{this.name}}</span>
+        <span>{{>manaCost}}</span>
+    </h4>
+    <h5>Oracle Text</h5>
+    <p class="card-type">{{this.type}}</p>
+    <p class="card-oracle-test">{{this.oracleText}}</p>
+    <p class="card-rarity">{{this.rarity}}</p>
+    <p class="card-artist">{{this.artist}}</p>
+    <p class="card-number">{{this.number}}</p>
+    <table class="card-legalities">
+        <tbody>
+        {{!-- TODO: hard code each format? I.e.: 
+            <td>Standard</td>
+            <td>{{this.standard}}</td>
+        etc --}}
+            {{#each this.legalities}}
+            <tr>
+                <td>{{this.key}}</td>
+                <td>{{this.value}}</td>
+            </tr>
+            {{/each}}
+        </tbody>
+    </table>
+</div>
+
+<div id="price-info"><!-- TODO: fill in with price info --></div>
+
+<div id="collection-info">
+    <div>
+        {{>cardsOwned}}
+    </div>
+    <div id="card-other-printings">
+        <h4 id="card-other-printings-title">Other Printings</h4>
+        <table>
+            <tbody>
+            {{#each this.otherPrintings}}
+                <tr>
+                    <td class="set-icon"><i class="ss ss-{{this.keyruneCode}}"></i></td>
+                    <td class="other-printing-url"><a href="{{this.url}}">{{this.set.name}}</a></td>
+                    <td class="other-differentiators">
+                    {{#each this.differentiators}}
+                        <span class="differentiators-description">{{this.description}}</span>
+                    {{/each}}
+                    </td>
+                    <td class="other-printing-price">{{this.price}}</td>
+                </tr>
+            {{/each}}
+            </tbody>
+        </table>
+    </div>
+</div>
