+++ test/http/views/cardPrintings.hbs
@@ -1,0 +1,42 @@
+{{!-- all instances of a single card --}}
+<div id="card-identity">
+    <h2>
+        <span id="card-mana-cost">{{>manaCost}}</span>
+        <span id="card-name">{{this.name}}</span>
+        <span> - All Printings</span>
+    </h2>
+    <table id="card-printings-metadata">
+        <tbody>
+            {{#each this.metadata}}
+            <tr>
+                <td class="metadata-key">{{this.key}}</td>
+                <td class="metadata-value">{{this.value}}</td>
+            </tr>
+            {{/each}}
+        </tbody>
+    </table>
+    <p class="card-oracle-text">{{this.oracleText}}</p>
+</div>
+<table id="all-printings-card-listing">
+    <tbody>
+    {{#each this.cards}}
+        <tr>
+            <td class="card-printing-owned">
+                {{>cardsOwned}}
+            </td>
+            <td class="card-printing-set">
+                <i class="ss ss-{{this.keyruneCode}} ss-fw"></i>
+                <span>{{this.set.name}}</span>
+            </td>
+            <td class="card-printing-icon">
+                <a href="{{this.url}}">
+                    <img src="{{this.image}}" alt="{{this.name}} ({{this.set.code}})" />
+                </a>
+            </td>
+            <td class="card-printing-price">
+                <p>{{this.price}}</p>
+            </td>
+        </tr>
+    {{/each}}
+    </tbody>
+</table>
