+++ test/http/views/error.hbs
@@ -1,0 +1,1 @@
+<div>Error.</div>
