+++ test/http/views/index.hbs
@@ -1,0 +1,3 @@
+<div>
+    {{>setList}}
+</div>
