+++ test/http/views/layouts/main.hbs
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>{{title}}</title>
+    <link href="//cdn.jsdelivr.net/npm/keyrune@latest/css/keyrune.css" rel="stylesheet" type="text/css" />
+</head>
+<body>
+    <header>
+        {{>navbar}}
+    </header>
+    <div>
+        {{{body}}}
+    </div>
+    <footer>
+        {{>footer}}
+    </footer>
+</body>
+</html>
