+++ test/http/views/partials/cardsOwned.hbs
@@ -1,0 +1,10 @@
+<span class="total-owned-controller">
+    <button class="remove-card-btn">
+        <span class="minus">-</span>
+    </button>
+    <span class="total-owned">{{#if this.totalOwned}}{{this.totalOwned}}{{else}}0{{/if}}</span>
+    <button class="add-card-btn">
+        <span class="plus">+</span>
+    </button>
+</span>
+{{!-- TODO: buttons for foil or list as entirely different card? --}}
