+++ test/http/views/partials/footer.hbs
@@ -1,0 +1,3 @@
+<div>
+    <h2>This is where the footer will go</h2>
+</div>
