+++ test/http/views/partials/manaCost.hbs
@@ -1,0 +1,5 @@
+<span class="mana-cost">
+    {{#each this.manaCost}}
+    <i class="ms ms-cost ms-shadow ms-{{this}}"></i>
+    {{/each}}
+</span>
