+++ test/http/views/partials/navbar.hbs
@@ -1,0 +1,9 @@
+<nav>
+    <ul>
+        <li><a href="/">Home</a></li>
+        <li><a href="/sets">Sets</a></li>
+        <li><a href="#">Collection</a></li>
+        <li><a href="#">Profile</a></li>
+        <li><a href="#">Settings</a></li>
+    </ul>
+</nav>
