+++ test/http/views/partials/setInfo.hbs
@@ -1,0 +1,11 @@
+<div id="set-info">
+    <h2 id="set-header">
+        <i class="ss ss-{{this.keyruneCode}} ss-2x"></i>
+        <span id="set-name">{{this.name}}</span>
+    </h2>
+    <ul id="set-metadata">
+        <li id="set-metadata-block">Block: {{this.block}}</li>
+        <li id="set-metadata-code">Code: {{this.code}}</li>
+        <li id="set-metadata-release-date">Release Date: {{this.releaseDate}}</li>
+    </ul>
+</div>
