+++ test/http/views/partials/setList.hbs
@@ -1,0 +1,14 @@
+<ul>
+    {{!-- TODO:
+    - in desc order 
+    - separate normal sets and special sets
+      - or toggle checkbox to show special sets?
+    - ? Group by legality/format ???
+    --}}
+    {{#each setList}}
+        <li class="set-name">
+            <i class="ss ss-{{this.keyruneCode}} ss-fw"></i>
+            <a href="{{this.url}}">{{this.name}}</a>
+        </li>
+    {{/each}}
+</ul>
