+++ test/http/views/profile.hbs
@@ -1,0 +1,10 @@
+<div id="user-profile-settings">
+    <h2 id="settings-header">Profile Settings</h2>
+    <form>
+        <label for="username">Username</label>
+        <input id="username" type="text" value="{{this.username}}"/>
+        <label for="user-email">Email</label>
+        <input id="user-email" value="{{this.email}}"/>
+        <input type="submit" value="Update Profile"/>
+    </form>
+</div>
