+++ test/http/views/set.hbs
@@ -1,0 +1,7 @@
+<link href="//cdn.jsdelivr.net/npm/mana-font@latest/css/mana.css" rel="stylesheet" type="text/css" />
+<div>
+    {{>setInfo}}
+</div>
+<div>
+    {{>setCards}}
+</div>
