+++ test/mtgjson-ingestion/mtgjson-ingestion-test-utils.ts
@@ -1,9 +1,9 @@
-import { CardSet } from './dtos/cardSet.dto';
-import { SetDto } from './dtos/set.dto';
-import { Identifiers } from './dtos/identifiers.dto';
-import { SetList } from './dtos/setList.dto';
-import { Set } from 'src/core/set/set';
-import { Card } from 'src/core/card/card';
+import { CardSet } from '../../src/mtgjson-ingestion/dtos/cardSet.dto'
+import { SetDto } from '../../src/mtgjson-ingestion/dtos/set.dto';
+import { Identifiers } from '../../src/mtgjson-ingestion/dtos/identifiers.dto';
+import { SetList } from '../../src/mtgjson-ingestion/dtos/setList.dto';
+import { Set } from '../../src/core/set/set';
+import { Card } from '../../src/core/card/card';
 
 export class MtgJsonIngestionTestUtils {
